requirements.txt
jupyter
azureml-sdk
numpy
matplotlib
onnx
tensorflow
